+++ packages/chrome-extension/src/cortex/events.ts
@@ -1,3 +1,0 @@
-import { event } from '@legendapp/state';
-
-export const scrollEventsToBottom = event();