+++ packages/chrome-extension/src/cortex/store/_store.ts
@@ -1,7 +1,0 @@
-import { chromeStore } from './chrome.store';
-import { windowsStore } from './windows.store';
-
-export const store = {
-  chrome: chromeStore,
-  windows: windowsStore,
-};