+++ packages/chrome-extension/src/cortex/store/windows.store.ts
@@ -1,7 +1,0 @@
-type WindowsStore = {
-  index: number;
-};
-
-export const windowsStore: WindowsStore = {
-  index: 0,
-};